+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/header/table-menu/o-table-menu.component.ts
@@ -351,7 +351,9 @@ export class OTableMenuComponent implements OTableMenu, OnInit, AfterViewInit, O
       data: {
         visibleColumns: Util.parseArray(this.table.visibleColumns, true),
         columnsData: this.table.oTableOptions.columns,
-        rowHeight: this.table.rowHeight
+        rowHeight: this.table.rowHeight,
+        activeColumnValueFilters: this.table.dataSource.getColumnValueFilters().map(colValueFilter => colValueFilter.attr),
+        activeSortColumns: this.table.sortColArray.map(col => col.columnName)
       },
       maxWidth: '35vw',
       disableClose: true,
@@ -363,6 +365,13 @@ export class OTableMenuComponent implements OTableMenu, OnInit, AfterViewInit, O
         this.table.visibleColArray = dialogRef.componentInstance.getVisibleColumns();
         const columnsOrder = dialogRef.componentInstance.getColumnsOrder();
         this.table.oTableOptions.columns.sort((a: OColumn, b: OColumn) => columnsOrder.indexOf(a.attr) - columnsOrder.indexOf(b.attr));
+        const columnValueFiltersToRemove = dialogRef.componentInstance.getColumnValueFiltersToRemove();
+        const columnSortingToRemove = dialogRef.componentInstance.getColumnSortingToRemove();
+        if (columnValueFiltersToRemove.length > 0 || columnSortingToRemove.length > 0) {
+          const sortColumns = this.table.sortColArray.filter(col => !columnSortingToRemove.includes(col.columnName))
+          this.table.reinitializeSortColumns(sortColumns);
+          this.table.clearColumnFilters(true, columnValueFiltersToRemove);
+        }
         this.table.cd.detectChanges();
         this.table.refreshColumnsWidth();
       }